Confirms deletion of runs
Just uses a browser alert feature

+++ Project/stats/run.php
@@ -123,7 +123,7 @@ class Run {
 						  <div class="panel-footer">&nbsp;
 						    <span class="pull-right">
 							  <a href="editRun.php?ID=' . $this->RunID . '" title="Edit" data-toggle="tooltip" type="button" class="btn btn-sm btn-warning"><i class="glyphicon glyphicon-edit"></i></a>
-							  <a href="deleteRun.php?ID=' . $this->RunID . '" title="Delete" data-toggle="tooltip" type="button" class="btn btn-sm btn-danger"><i class="glyphicon glyphicon-trash"></i></a>
+							  <a href="deleteRun.php?ID=' . $this->RunID . '" title="Delete" data-toggle="tooltip" type="button" class="btn btn-sm btn-danger" onclick="return confirm(\'Are you sure you want to delete this item?\');"><i class="glyphicon glyphicon-trash"></i></a>
 							</span>
 						  </div>
 						</div>';
